Fix draggable device switcher
