+++ react/Redirects.tsx
@@ -3,7 +3,7 @@ import React, { Component } from 'react'
 export default class Redirects extends Component {
   public render() {
     return (
-      <p>test</p>
+      null
     )
   }
 }
